backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
python-dotenv>=1.0.0
aiofiles>=23.1.0
python-multipart>=0.0.6
pypdf>=3.12.0
pdfplumber>=0.10.3
pdf2image>=1.17.0
Pillow>=10.0.0
reportlab>=4.0.0
python-docx>=1.0.0
pytesseract>=0.3.10
openpyxl>=3.1.2
boto3>=1.34.0
botocore>=1.34.0
